backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai
pydantic
pydantic[email]
python-jose[cryptography]
fpdf2